% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_browser.R
\name{browser_css}
\alias{browser_css}
\title{CSS for Browser Interface}
\usage{
browser_css()
}
\value{
HTML head tag with CSS
}
\description{
Returns CSS styling for browser components.
}
\keyword{internal}
