% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{check_input_directory}
\alias{check_input_directory}
\title{Check Input Directory Structure}
\usage{
check_input_directory(.dir)
}
\arguments{
\item{.dir}{Path to classification project directory}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Check Input Directory Structure
}
