% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_all_marks}
\alias{clear_all_marks}
\title{Clear All Marks}
\usage{
clear_all_marks(marked_docs)
}
\arguments{
\item{marked_docs}{ReactiveValues object with 'ids' component}
}
\value{
Integer count of cleared marks
}
\description{
Clears all marked documents.
}
