% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{delete_export}
\alias{delete_export}
\title{Delete Export}
\usage{
delete_export(.dir, .filename)
}
\arguments{
\item{.dir}{Path to project directory}

\item{.filename}{Export filename to delete}
}
\value{
Logical TRUE if successful
}
\description{
Deletes an export file.
}
