% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{filter_documents_by_classification}
\alias{filter_documents_by_classification}
\title{Filter Documents by Classification Criteria}
\usage{
filter_documents_by_classification(.dir, .class_filters)
}
\arguments{
\item{.dir}{Path to classification project directory}

\item{.class_filters}{Named list of classification filters (class_name -> selected_value)}
}
\value{
Vector of document IDs that match the criteria
}
\description{
Filter Documents by Classification Criteria
}
