% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generate_multiselect_warning}
\alias{generate_multiselect_warning}
\title{Generate Multi-Selection Warning}
\usage{
generate_multiselect_warning(selections_nested, schema, mode)
}
\arguments{
\item{selections_nested}{Nested list of current selections}

\item{schema}{Schema list}

\item{mode}{Current selection mode}
}
\value{
Shiny UI element or NULL
}
\description{
Creates warning badge when in single mode but multiple values selected.
}
