% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generate_schema_tab_badge}
\alias{generate_schema_tab_badge}
\title{Generate Schema Tab Badge}
\usage{
generate_schema_tab_badge(
  schema_id,
  schema_name,
  completed_classes,
  total_classes
)
}
\arguments{
\item{schema_id}{Schema ID}

\item{schema_name}{Schema name}

\item{completed_classes}{Number of completed classes}

\item{total_classes}{Total number of classes}
}
\value{
Character string for tab title
}
\description{
Creates completion badge for schema tab label.
}
\keyword{internal}
