% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_classification_stats}
\alias{get_classification_stats}
\title{Get Classification Statistics}
\usage{
get_classification_stats(.dir)
}
\arguments{
\item{.dir}{Path to classification project directory}
}
\value{
Data frame with classification statistics
}
\description{
Get Classification Statistics
}
