% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Get Database Connection}
\usage{
get_db_connection(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
DBI connection object
}
\description{
Helper function to get SQLite connection.
}
\keyword{internal}
