% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{get_docids}
\alias{get_docids}
\title{Get All Document IDs}
\usage{
get_docids(.dir, .type = c("All", "Classified", "Unclassified"))
}
\arguments{
\item{.dir}{Path to project directory}

\item{.type}{Type filter: "All", "Classified", or "Unclassified"}
}
\value{
Character vector of document IDs
}
\description{
Retrieves list of document IDs, optionally filtered by classification status.
}
