% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{get_docids_with_notes}
\alias{get_docids_with_notes}
\title{Get Document IDs with Notes}
\usage{
get_docids_with_notes(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
Character vector of document IDs
}
\description{
Returns list of document IDs that have notes (latest note is not NULL).
}
