% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{get_documents_db_connection}
\alias{get_documents_db_connection}
\title{Get Documents Database Connection}
\usage{
get_documents_db_connection(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
DBI connection object
}
\description{
Helper function to get SQLite connection for documents.
}
\keyword{internal}
