% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_progress_stats}
\alias{get_progress_stats}
\title{Get Classification Progress Statistics}
\usage{
get_progress_stats(.dir)
}
\arguments{
\item{.dir}{Path to classification project directory}
}
\value{
List with progress statistics
}
\description{
Get Classification Progress Statistics
}
