% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{get_schema_tab_id}
\alias{get_schema_tab_id}
\title{Get Schema Tab ID}
\usage{
get_schema_tab_id(schema_id)
}
\arguments{
\item{schema_id}{Schema ID}
}
\value{
Character tab ID
}
\description{
Generates consistent tab ID for a schema.
}
\keyword{internal}
