% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{initialize_classification_file}
\alias{initialize_classification_file}
\title{Initialize Classification File}
\usage{
initialize_classification_file(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
Invisible NULL
}
\description{
Creates empty ClassificationDetails.parquet if it doesn't exist.
}
\keyword{internal}
