% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_logic.R
\name{is_document_classified}
\alias{is_document_classified}
\title{Check if Document is Classified}
\usage{
is_document_classified(.dir, .doc_id)
}
\arguments{
\item{.dir}{Path to project directory}

\item{.doc_id}{Document ID to check}
}
\value{
Logical TRUE if document has any classifications
}
\description{
Determines if a document has any classifications.
}
