% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{list_exports}
\alias{list_exports}
\title{List Exports}
\usage{
list_exports(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
Data frame with export file information
}
\description{
Lists all export files in the exports directory.
}
