% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_schema_class_key}
\alias{make_schema_class_key}
\title{Generate schema-class key}
\usage{
make_schema_class_key(schema_id, class_name)
}
\arguments{
\item{schema_id}{Schema ID}

\item{class_name}{Class name}
}
\value{
Character key in format "schema_class"
}
\description{
Generate schema-class key
}
\keyword{internal}
