% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mark_documents}
\alias{mark_documents}
\title{Mark Documents}
\usage{
mark_documents(marked_docs, doc_ids)
}
\arguments{
\item{marked_docs}{ReactiveValues object with 'ids' component}

\item{doc_ids}{Character vector of document IDs to mark}
}
\value{
Invisible NULL
}
\description{
Adds document IDs to the marked documents list.
}
