% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{mod_browser_server}
\alias{mod_browser_server}
\title{Browser Module Server}
\usage{
mod_browser_server(id, .dir, schema)
}
\arguments{
\item{id}{Module namespace ID}

\item{.dir}{Path to classification directory}

\item{schema}{Classification schema}
}
\description{
Browser Module Server
}
