% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_browser.R
\name{mod_browser_ui}
\alias{mod_browser_ui}
\title{Browser Module UI}
\usage{
mod_browser_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI
}
\description{
Generates the user interface for the browser/filtering tab.
}
