% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_classification.R
\name{mod_classification_server}
\alias{mod_classification_server}
\title{Classification Module Server}
\usage{
mod_classification_server(id, .dir, .user_id, schema)
}
\arguments{
\item{id}{Module namespace ID}

\item{.dir}{Reactive or static path to project directory}

\item{.user_id}{Reactive or static user ID}

\item{schema}{Reactive or static schema list}
}
\description{
Server logic for the classification module.
}
