% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_classification.R
\name{mod_classification_ui}
\alias{mod_classification_ui}
\title{Classification Module UI}
\usage{
mod_classification_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI
}
\description{
Generates the user interface for the classification tab.
}
