% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export.R
\name{mod_export_server}
\alias{mod_export_server}
\title{Export Module Server}
\usage{
mod_export_server(id, .dir, schema, marked_docs = NULL)
}
\arguments{
\item{id}{Module namespace ID}

\item{.dir}{Reactive or static path to project directory}

\item{schema}{Reactive or static schema list}

\item{marked_docs}{ReactiveValues object with marked document IDs}
}
\description{
Server logic for the export module.
}
