% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro.R
\name{mod_intro_server}
\alias{mod_intro_server}
\title{Introduction Module Server}
\usage{
mod_intro_server(id, .dir)
}
\arguments{
\item{id}{Module namespace ID}

\item{.dir}{Reactive or static path to project directory}
}
\description{
Server logic for the introduction module.
}
