% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro.R
\name{mod_intro_ui}
\alias{mod_intro_ui}
\title{Introduction Module UI}
\usage{
mod_intro_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI
}
\description{
Generates the user interface for the introduction/setup tab.
Shows upload interface if files missing, or data overview if ready.
}
