% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{mod_overview_ui}
\alias{mod_overview_ui}
\title{Overview Module UI}
\usage{
mod_overview_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI
}
\description{
Generates the user interface for the overview/analytics tab.
}
