% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{overview_css}
\alias{overview_css}
\title{CSS for Overview Interface}
\usage{
overview_css()
}
\value{
HTML head tag with CSS
}
\description{
Returns CSS styling for overview components.
}
\keyword{internal}
