% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_schema_class_key}
\alias{parse_schema_class_key}
\title{Parse schema-class key}
\usage{
parse_schema_class_key(key)
}
\arguments{
\item{key}{Character key in format "schema_class"}
}
\value{
List with schema_id and class_name
}
\description{
Parse schema-class key
}
\keyword{internal}
