% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{plot_classification_distribution}
\alias{plot_classification_distribution}
\title{Create Classification Distribution Plot}
\usage{
plot_classification_distribution(.dir, .class_name)
}
\arguments{
\item{.dir}{Path to classification project directory}

\item{.class_name}{Classification class to plot}
}
\value{
ggplot object showing distribution of values for a classification class
}
\description{
Create Classification Distribution Plot
}
