% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{plot_progress_overview}
\alias{plot_progress_overview}
\title{Create Progress Overview Plot}
\usage{
plot_progress_overview(.dir)
}
\arguments{
\item{.dir}{Path to classification project directory}
}
\value{
ggplot object showing classification progress
}
\description{
Create Progress Overview Plot
}
