% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{plot_user_activity}
\alias{plot_user_activity}
\title{Create User Activity Plot}
\usage{
plot_user_activity(.dir)
}
\arguments{
\item{.dir}{Path to classification project directory}
}
\value{
ggplot object showing activity by user
}
\description{
Create User Activity Plot
}
