% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{read_all_classifications}
\alias{read_all_classifications}
\title{Read All Classifications}
\usage{
read_all_classifications(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
Data frame with all classification records
}
\description{
Loads all classifications from the database.
}
