% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{read_all_notes}
\alias{read_all_notes}
\title{Read All Notes}
\usage{
read_all_notes(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
Data frame with all latest note records
}
\description{
Loads the latest notes for all documents from SQLite.
}
