% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{read_classification}
\alias{read_classification}
\title{Read Classifications}
\usage{
read_classification(.dir, .doc_id = NULL)
}
\arguments{
\item{.dir}{Path to classification project directory}

\item{.doc_id}{Optional document ID to filter to specific document}
}
\value{
Data frame with classifications
}
\description{
Read Classifications
}
