% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{read_schema}
\alias{read_schema}
\title{Read Classification Schema}
\usage{
read_schema(.dir)
}
\arguments{
\item{.dir}{Path to classification project directory}
}
\value{
List with schema (Class -> Values)
}
\description{
Read Classification Schema
}
