% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro.R
\name{render_help_section}
\alias{render_help_section}
\title{Render Help Section}
\usage{
render_help_section()
}
\value{
Shiny UI
}
\description{
Creates comprehensive help documentation.
}
\keyword{internal}
