% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro.R
\name{render_overview_interface}
\alias{render_overview_interface}
\title{Render Overview Interface}
\usage{
render_overview_interface(ns, dir_path)
}
\arguments{
\item{ns}{Namespace function}

\item{dir_path}{Directory path (character)}
}
\value{
Shiny UI
}
\description{
Shows project overview and schema when files are ready.
}
\keyword{internal}
