% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro.R
\name{render_schema_display}
\alias{render_schema_display}
\title{Render Schema Display}
\usage{
render_schema_display(schema)
}
\arguments{
\item{schema}{Schema list}
}
\value{
Shiny UI
}
\description{
Creates nicely formatted display of schema structure.
}
\keyword{internal}
