% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro.R
\name{render_setup_interface}
\alias{render_setup_interface}
\title{Render Setup Interface}
\usage{
render_setup_interface(ns, status, dir_path)
}
\arguments{
\item{ns}{Namespace function}

\item{status}{Setup status list}

\item{dir_path}{Directory path (character)}
}
\value{
Shiny UI
}
\description{
Shows file upload interface when project is not ready.
}
\keyword{internal}
