% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{save_classification}
\alias{save_classification}
\title{Save Classification}
\usage{
save_classification(.dir, .doc_id, .user_id, .classifications_list)
}
\arguments{
\item{.dir}{Path to project directory}

\item{.doc_id}{Document ID being classified}

\item{.user_id}{User making the classification}

\item{.classifications_list}{Nested list structure}
}
\value{
Invisible NULL
}
\description{
Saves classification values for a document in SQLite append-only log.
}
