% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_classification}
\alias{start_classification}
\title{Quick Start Function}
\usage{
start_classification(project_dir, user = Sys.getenv("USER"))
}
\arguments{
\item{project_dir}{Path to project directory}

\item{user}{User identifier}
}
\value{
Shiny app object
}
\description{
Simplified wrapper for common use case.
}
