% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmark_documents}
\alias{unmark_documents}
\title{Unmark Documents}
\usage{
unmark_documents(marked_docs, doc_ids)
}
\arguments{
\item{marked_docs}{ReactiveValues object with 'ids' component}

\item{doc_ids}{Character vector of document IDs to unmark}
}
\value{
Invisible NULL
}
\description{
Removes document IDs from the marked documents list.
}
