% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{update_schema_hash}
\alias{update_schema_hash}
\title{Update Schema Hash}
\usage{
update_schema_hash(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
Invisible NULL
}
\description{
Updates stored schema hash after user confirms schema change.
}
