% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_schema_file}
\alias{validate_schema_file}
\title{Validate Schema File}
\usage{
validate_schema_file(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Checks schema file for required structure and valid values.
}
\keyword{internal}
