% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{with_documents_db_connection}
\alias{with_documents_db_connection}
\title{Execute Function with Documents Database Connection}
\usage{
with_documents_db_connection(.dir, .func)
}
\arguments{
\item{.dir}{Path to project directory}

\item{.func}{Function to execute with connection}
}
\value{
Result of .func
}
\description{
Wrapper that handles documents database connection lifecycle.
}
\keyword{internal}
