% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{classification_app}
\alias{classification_app}
\title{Main Classification App with Tabs}
\usage{
classification_app(.dir, .user_id = "user")
}
\arguments{
\item{.dir}{Path to classification directory}

\item{.user_id}{User ID}
}
\description{
Main Classification App with Tabs
}
