% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{format_document_info}
\alias{format_document_info}
\title{Format Document Info Display}
\usage{
format_document_info(
  .dir,
  .doc_id,
  .current_index,
  .total_filtered,
  .filter_type,
  .schema
)
}
\arguments{
\item{.dir}{Path to classification project directory}

\item{.doc_id}{Current document ID (can be NULL)}

\item{.current_index}{Current position in filtered list}

\item{.total_filtered}{Total number of filtered documents}

\item{.filter_type}{Current filter type}

\item{.schema}{Classification schema}
}
\value{
Formatted string for document info display
}
\description{
Format Document Info Display
}
