% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generate_classification_dropdown}
\alias{generate_classification_dropdown}
\title{Generate Classification Dropdown}
\usage{
generate_classification_dropdown(
  ns,
  schema_id,
  class_name,
  values,
  selected_values,
  mode = "multi"
)
}
\arguments{
\item{ns}{Namespace function from module}

\item{schema_id}{Schema ID}

\item{class_name}{Class name}

\item{values}{Available values for this class}

\item{selected_values}{Currently selected values}

\item{mode}{"single" or "multi"}
}
\value{
Shiny input element (selectInput or selectizeInput)
}
\description{
Creates a classification dropdown (single or multi-select based on mode).
}
\keyword{internal}
