% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_docids}
\alias{get_docids}
\title{Get All Document IDs (for navigation)}
\usage{
get_docids(.dir, .type = c("All", "Classified", "Unclassified"))
}
\arguments{
\item{.dir}{Path to classification project directory}

\item{.type}{Type of documents to return}
}
\value{
Vector of document IDs
}
\description{
Get All Document IDs (for navigation)
}
