% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_document_count}
\alias{get_document_count}
\title{Get Document Count (without loading all data)}
\usage{
get_document_count(.dir)
}
\arguments{
\item{.dir}{Path to classification project directory}
}
\value{
Number of documents
}
\description{
Get Document Count (without loading all data)
}
