% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{get_document_count}
\alias{get_document_count}
\title{Get Document Count}
\usage{
get_document_count(.dir)
}
\arguments{
\item{.dir}{Path to project directory}
}
\value{
Integer count of documents
}
\description{
Returns total number of documents.
}
