% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_filter_counts}
\alias{get_filter_counts}
\title{Get Classification Count by Filter Type}
\usage{
get_filter_counts(.dir)
}
\arguments{
\item{.dir}{Path to classification directory}
}
\value{
Named list with counts for each filter type
}
\description{
Get Classification Count by Filter Type
}
