% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{is_document_classified}
\alias{is_document_classified}
\title{Check if Document is Fully Classified}
\usage{
is_document_classified(.dir, .doc_id)
}
\arguments{
\item{.dir}{Path to classification directory}

\item{.doc_id}{Document ID to check}
}
\value{
TRUE if document has all required classifications
}
\description{
Check if Document is Fully Classified
}
