% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{mod_classification_server}
\alias{mod_classification_server}
\title{Classification Module Server}
\usage{
mod_classification_server(id, .dir, .user_id, schema)
}
\arguments{
\item{id}{Module namespace ID}

\item{.dir}{Path to classification directory}

\item{.user_id}{User ID}

\item{schema}{Classification schema}
}
\description{
Classification Module Server
}
