% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{mod_overview_server}
\alias{mod_overview_server}
\title{Overview Module Server (FIXED VERSION)}
\usage{
mod_overview_server(id, .dir)
}
\arguments{
\item{id}{Module namespace ID}

\item{.dir}{Path to classification directory}
}
\description{
Overview Module Server (FIXED VERSION)
}
