% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{read_single_document}
\alias{read_single_document}
\title{Read Single Document by ID}
\usage{
read_single_document(.dir, .doc_id)
}
\arguments{
\item{.dir}{Path to classification project directory}

\item{.doc_id}{Document ID to retrieve}
}
\value{
Data frame with single document (DocID, HTML columns) or
empty data frame if document not found
}
\description{
Read Single Document by ID
}
