% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{read_single_document}
\alias{read_single_document}
\title{Read Single Document}
\usage{
read_single_document(.dir, .doc_id)
}
\arguments{
\item{.dir}{Path to project directory}

\item{.doc_id}{Document ID to retrieve}
}
\value{
Data frame with DocID and HTML columns
}
\description{
Retrieves a single document by DocID from SQLite.
}
