% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{save_classification}
\alias{save_classification}
\title{Save Classification}
\usage{
save_classification(.dir, .doc_id, .user_id, .classifications)
}
\arguments{
\item{.dir}{Path to classification project directory}

\item{.doc_id}{Document ID}

\item{.user_id}{User ID}

\item{.classifications}{Named list of classifications}
}
\description{
Save Classification
}
